% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency.R
\name{html_dependency_tacotable}
\alias{html_dependency_tacotable}
\title{Dependencies for react-taco-table}
\usage{
html_dependency_tacotable()
}
\value{
\code{htmlDependency}
}
\description{
Dependencies for react-taco-table
}

